[Refactor] Make casts of enum class to int explicit in fmt::format(). (#25806)
The old fmt-7.0.3 being used in Starrocks converted enum class to int
implicitly.
But it is not allowed anymore in the recent fmt library and will cause a
compile error.
So this explicit cast is a prerequisite for fmt library upgrade which
I'm planning to request in a separate PR.

(cherry picked from commit b092a91bbdd3042afa69113be92a5fe33090f4ed)

+++ be/src/io/s3_input_stream.cpp
@@ -28,8 +28,9 @@
 namespace starrocks::io {
 
 inline Status make_error_status(const Aws::S3::S3Error& error) {
-    return Status::IOError(fmt::format("code={}(SdkErrorType:{}), message={}", error.GetResponseCode(),
-                                       error.GetErrorType(), error.GetMessage()));
+    return Status::IOError(fmt::format("code={}(SdkErrorType:{}), message={}",
+                                       static_cast<int>(error.GetResponseCode()),
+                                       static_cast<int>(error.GetErrorType()), error.GetMessage()));
 }
 
 StatusOr<int64_t> S3InputStream::read(void* out, int64_t count) {
